requirements.txt
openai
pyyaml
pillow
matplotlib